Feat/reorganise templates (#14)
* reorganise files in template folders

* feat: use submodule exports
