+++ src/index.ts
@@ -1,2 +1,1 @@
-export * from "./classic/components"
 export * from "./engine"
