+++ src/templates/classic/components/Content/Content.tsx
