+++ src/templates/classic/components/Content/index.ts
