+++ src/templates/classic/components/Footer/Footer.tsx
