+++ src/templates/classic/components/Footer/index.ts
