+++ src/templates/classic/components/Hero/Hero.stories.tsx
