+++ src/templates/classic/components/Hero/Hero.tsx
