+++ src/templates/classic/components/Hero/HeroInfobox.tsx
