+++ src/templates/classic/components/Hero/index.ts
