+++ src/templates/classic/components/Image/Image.stories.tsx
