+++ src/templates/classic/components/Image/Image.tsx
