+++ src/templates/classic/components/Image/index.ts
