+++ src/templates/classic/components/InfoCards/InfoCards.tsx
