+++ src/templates/classic/components/InfoCards/index.ts
