+++ src/templates/classic/components/Infobar/Infobar.tsx
