+++ src/templates/classic/components/Infobar/index.ts
