+++ src/templates/classic/components/Infopic/Infopic.stories.tsx
