+++ src/templates/classic/components/Infopic/Infopic.tsx
