+++ src/templates/classic/components/Infopic/index.ts
