+++ src/templates/classic/components/Masthead/Masthead.stories.tsx
