+++ src/templates/classic/components/Masthead/Masthead.tsx
