+++ src/templates/classic/components/Masthead/index.ts
