+++ src/templates/classic/components/Search/Search.tsx
