+++ src/templates/classic/components/Search/index.tsx
