+++ src/templates/classic/components/SidePane/SidePane.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryFn } from "@storybook/react"
 import SidePane from "./SidePane"
-import Sitemap from "../../../sitemap.json"
+import Sitemap from "../../../../sitemap.json"
 import { SidePaneProps } from "~/common"
 export default {
   title: "Classic/Components/SidePane",
