+++ src/templates/classic/components/SidePane/SidePane.tsx
