+++ src/templates/classic/components/SidePane/index.ts
