+++ src/templates/classic/components/index.ts
