+++ src/templates/classic/index.ts
