+++ src/templates/next/components/Navbar/Navbar.stories.tsx
