+++ src/templates/next/components/index.ts
@@ -1,1 +1,2 @@
 export { default as Button } from "./Button"
+export { default as Navbar } from "./Navbar"
