+++ src/templates/next/index.ts
